Update outputDatasetTag
Add "-130X_DeepMET" to the end of the output tag to more closely match
the standard dataset convention. This matches the style of global tags
in centrally produced datasets. The datasets have already been produced
without this added naming convention, and this is more to serve as a
model for future jobs.